backend/churn_prediction/churn_prediction/requirements.txt
django
pymysql
scikit-learn
joblib
numpy==2.0.2
djangorestframework